move search and filter icons
